renamed RightPanel as Tab
